Refactor : UI Pipeline details page
